chore: Add some flags to compiler
Part of #43

Add `-Werror -Wextra -Wpedantic`

For main lib it was not possible to add -Werror as YesChief! contains
some warning that may be fixed in future release

+++ tests/e2e/test_run.test.cpp
@@ -32,7 +32,7 @@ describe(test_run, []() {
     it("Should have the same count of test between list and run", []() {
         const auto list      = run_with_args("list");
         const auto test_list = split(list, "\n");
-        assertThat(test_list.size()).isGreaterThanOrEqualTo(1);
+        assertThat(test_list.size()).isGreaterThanOrEqualTo(static_cast<unsigned int>(1));
 
         const auto run = run_with_args("run");
         assertThat(run).hasSubString("TEST COUNT: " + std::to_string(test_list.size()));
